google_sheets.py
from google.oauth2 import service_account
from googleapiclient.discovery import build

class SheetService:
    def __init__(self, sheet_id, sheet_name):
        self.sheet_id = sheet_id
        self.sheet_name = sheet_name

    def append_row(self, data):
        # Implement Google Sheets API append logic here
        print(f"Appending to sheet: {data}")
